- split pools into separate files - converted committed pool lock to RWMutex - added RLock in get()
