+++ services/transactionpool/pending_pool.go
@@ -18,13 +18,6 @@ func NewPendingPool(pendingPoolSizeInBytes func() uint32) *pendingTxPool {
 	}
 }
 
-func NewCommittedPool() *committedTxPool {
-	return &committedTxPool{
-		transactions: make(map[string]*committedTransaction),
-		lock:         &sync.Mutex{},
-	}
-}
-
 type pendingTransaction struct {
 	gatewayPublicKey primitives.Ed25519PublicKey
 	transaction      *protocol.SignedTransaction
@@ -42,9 +35,6 @@ type pendingTxPool struct {
 }
 
 func (p *pendingTxPool) add(transaction *protocol.SignedTransaction, gatewayPublicKey primitives.Ed25519PublicKey) (primitives.Sha256, *ErrTransactionRejected) {
-	p.lock.Lock()
-	defer p.lock.Unlock()
-
 	size := sizeOf(transaction)
 
 	if p.currentSizeInBytes+size > p.pendingPoolSizeInBytes() {
@@ -53,6 +43,8 @@ func (p *pendingTxPool) add(transaction *protocol.SignedTransaction, gatewayPubl
 
 	key := digest.CalcTxHash(transaction.Transaction())
 
+	p.lock.Lock()
+	defer p.lock.Unlock()
 	if _, exists := p.transactionsByHash[key.KeyForMap()]; exists {
 		return nil, &ErrTransactionRejected{protocol.TRANSACTION_STATUS_DUPLICATE_TRANSACTION_ALREADY_PENDING}
 	}
@@ -66,6 +58,7 @@ func (p *pendingTxPool) add(transaction *protocol.SignedTransaction, gatewayPubl
 
 	return key, nil
 }
+
 func (p *pendingTxPool) has(transaction *protocol.SignedTransaction) bool {
 	p.lock.RLock()
 	defer p.lock.RUnlock()
@@ -77,6 +70,7 @@ func (p *pendingTxPool) has(transaction *protocol.SignedTransaction) bool {
 func (p *pendingTxPool) remove(txhash primitives.Sha256) *pendingTransaction {
 	p.lock.Lock()
 	defer p.lock.Unlock()
+
 	pendingTx, ok := p.transactionsByHash[txhash.KeyForMap()]
 	if ok {
 		delete(p.transactionsByHash, txhash.KeyForMap())
@@ -89,6 +83,8 @@ func (p *pendingTxPool) remove(txhash primitives.Sha256) *pendingTransaction {
 }
 
 func (p *pendingTxPool) getBatch(maxNumOfTransactions uint32, sizeLimitInBytes uint32) Transactions {
+
+	// this is not locked on purpose, the reasoning is that in the worst case scenario, we will iterate over another element that's just been added to the front of the list
 	txs := make(Transactions, 0, maxNumOfTransactions)
 	accumulatedSize := uint32(0)
 	e := p.transactionList.Back()
@@ -104,7 +100,7 @@ func (p *pendingTxPool) getBatch(maxNumOfTransactions uint32, sizeLimitInBytes u
 		tx := e.Value.(*protocol.SignedTransaction)
 		//
 		accumulatedSize += sizeOf(tx)
-		if uint32(len(txs)) >= maxNumOfTransactions || (sizeLimitInBytes > 0 && accumulatedSize > sizeLimitInBytes) {
+		if sizeLimitInBytes > 0 && accumulatedSize > sizeLimitInBytes {
 			break
 		}
 
@@ -117,6 +113,9 @@ func (p *pendingTxPool) getBatch(maxNumOfTransactions uint32, sizeLimitInBytes u
 }
 
 func (p *pendingTxPool) get(txHash primitives.Sha256) *protocol.SignedTransaction {
+	p.lock.RLock()
+	defer p.lock.RUnlock()
+
 	if ptx, ok := p.transactionsByHash[txHash.KeyForMap()]; ok {
 		return ptx.transaction
 	}
@@ -142,49 +141,6 @@ func (p *pendingTxPool) clearTransactionsOlderThan(time time.Time) {
 			p.remove(digest.CalcTxHash(tx.Transaction()))
 		}
 	}
-}
-
-type committedTxPool struct {
-	transactions map[string]*committedTransaction
-	lock         *sync.Mutex
-}
-
-func (p *committedTxPool) add(receipt *protocol.TransactionReceipt, ts primitives.TimestampNano) {
-	p.lock.Lock()
-	defer p.lock.Unlock()
-	p.transactions[receipt.Txhash().KeyForMap()] = &committedTransaction{
-		receipt:   receipt,
-		timestamp: ts,
-	}
-}
-
-func (p *committedTxPool) get(txHash primitives.Sha256) *committedTransaction {
-	key := txHash.KeyForMap()
-
-	tx := p.transactions[key]
-
-	return tx
-}
-
-func (p *committedTxPool) has(txHash primitives.Sha256) bool {
-	_, ok := p.transactions[txHash.KeyForMap()]
-	return ok
-}
-
-func (p *committedTxPool) clearTransactionsOlderThan(time time.Time) {
-	p.lock.Lock()
-	defer p.lock.Unlock()
-
-	for _, tx := range p.transactions {
-		if int64(tx.timestamp) < time.UnixNano() {
-			delete(p.transactions, tx.receipt.Txhash().KeyForMap())
-		}
-	}
-}
-
-type committedTransaction struct {
-	receipt   *protocol.TransactionReceipt
-	timestamp primitives.TimestampNano
 }
 
 func sizeOf(transaction *protocol.SignedTransaction) uint32 {
